+++ core/engine/src/test/java/com/gps19/core/engine/ForensicIdentityTest.kt
@@ -11,10 +11,10 @@ class ForensicIdentityTest {
     }
 
     private val listener = object : LocationProcessorListener {
-        override fun onTrailPointSaved(lat: Double, lng: Double, isViewerTrail: Boolean, isJump: Boolean, timestamp: Long, isHindsightCorrected: Boolean, accuracy: Float, maxAccuracy: Float) {}
-        override fun onLogAdded(message: String, type: String, isImportant: Boolean, isSpecial: Boolean, lat: Double, lng: Double, accuracy: Float, snr: Float?, vibe: Float?) {}
-        override fun onMaxAccuracyChanged(accuracy: Float) {}
-        override fun onChairBaselineChanged(baseline: Float) {}
+        override fun onTrailPointSaved(lat: Double, lng: Double, isViewerTrail: Boolean, isJump: Boolean, timestamp: Long, isHindsightCorrected: Boolean, accuracy: Double, maxAccuracy: Double) {}
+        override fun onLogAdded(message: String, type: String, isImportant: Boolean, isSpecial: Boolean, lat: Double, lng: Double, accuracy: Double, snr: Double?, vibe: Double?) {}
+        override fun onMaxAccuracyChanged(accuracy: Double) {}
+        override fun onChairBaselineChanged(baseline: Double) {}
         override fun onGpsStallDetected(ts: Long) {}
     }
 
@@ -24,7 +24,7 @@ class ForensicIdentityTest {
     fun `processGpsPoint handles valid point`() {
         val result = processor.processGpsPoint(
             lat = 32.0, lng = 34.0, alt = 10.0, androidSpeedKph = 0.0,
-            gpsTs = 1000L, accuracy = 10.0f, bearing = 0.0f, snr = 40.0f, satsUsed = 10,
+            gpsTs = 1000L, accuracy = 10.0, bearing = 0.0, snr = 40.0, satsUsed = 10,
             isViewerTrail = false, lastGpsTs = 0L, isLocal = true
         )
 
@@ -36,7 +36,7 @@ class ForensicIdentityTest {
         // First point to establish baseline
         processor.processGpsPoint(
             lat = 32.0, lng = 34.0, alt = 10.0, androidSpeedKph = 0.0,
-            gpsTs = 1000L, accuracy = 10.0f, bearing = 0.0f, snr = 40.0f, satsUsed = 10,
+            gpsTs = 1000L, accuracy = 10.0, bearing = 0.0, snr = 40.0, satsUsed = 10,
             isViewerTrail = false, lastGpsTs = 0L, isLocal = true
         )
 
@@ -44,7 +44,7 @@ class ForensicIdentityTest {
         // Note: Extreme single-point jumps are classified as OUTLIER by the current Sentinel logic
         val jumpResult = processor.processGpsPoint(
             lat = 40.0, lng = 50.0, alt = 10.0, androidSpeedKph = 0.0,
-            gpsTs = 2000L, accuracy = 10.0f, bearing = 0.0f, snr = 40.0f, satsUsed = 10,
+            gpsTs = 2000L, accuracy = 10.0, bearing = 0.0, snr = 40.0, satsUsed = 10,
             isViewerTrail = false, lastGpsTs = 1000L, isLocal = true
         )
